♻️  refactor: dto 디렉터리 구조 변경
챌린저용과 관리자용을 구분하기 위해 admin / common 으로 디렉터리 구조를 나눴습니다
