+++ src/main/java/umc/product/domain/checklist/service/ChecklistCommandServiceImpl.java
@@ -7,8 +7,8 @@ import org.springframework.transaction.annotation.Transactional;
 import umc.product.domain.checklist.entity.ChecklistMemberAnswer;
 import umc.product.domain.checklist.entity.enums.ChecklistType;
 import umc.product.domain.checklist.repository.ChecklistMemberAnswerRepository;
-import umc.product.domain.study.dto.request.StudyChecklistListRequest;
-import umc.product.domain.study.dto.response.StudyCommonResponse;
+import umc.product.domain.study.dto.common.request.StudyChecklistListRequest;
+import umc.product.domain.study.dto.common.response.StudyCommonResponse;
 import umc.product.domain.study.entity.StudyMember;
 import umc.product.domain.study.status.StudyErrorStatus;
 import umc.product.global.common.exception.RestApiException;
