+++ src/main/java/umc/product/domain/study/dto/common/request/StudyModifyRequest.java
@@ -1,4 +1,4 @@
-package umc.product.domain.study.dto.request;
+package umc.product.domain.study.dto.common.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
