+++ src/main/java/umc/product/domain/study/repository/StudyCustomRepository.java
@@ -1,8 +1,8 @@
 package umc.product.domain.study.repository;
 
-import umc.product.domain.study.dto.response.StudyMemberResponse;
-import umc.product.domain.study.dto.response.StudyWeekChecklistResponse;
-import umc.product.domain.study.dto.response.StudyWorkbookResponse;
+import umc.product.domain.study.dto.common.response.StudyMemberResponse;
+import umc.product.domain.study.dto.common.response.StudyWeekChecklistResponse;
+import umc.product.domain.study.dto.common.response.StudyWorkbookResponse;
 import umc.product.domain.study.entity.Study;
 import umc.product.domain.study.entity.StudyMember;
