+++ src/main/java/umc/product/domain/study/repository/StudyCustomRepositoryImpl.java
@@ -14,6 +14,9 @@ import umc.product.domain.member.entity.QMember;
 import umc.product.domain.roadmap.entity.QRoadmap;
 import umc.product.domain.roadmap.entity.QRoadmapSemester;
 import umc.product.domain.semester.entity.QSemesterPart;
+import umc.product.domain.study.dto.common.response.StudyMemberResponse;
+import umc.product.domain.study.dto.common.response.StudyWeekChecklistResponse;
+import umc.product.domain.study.dto.common.response.StudyWorkbookResponse;
 import umc.product.domain.study.dto.response.*;
 import umc.product.domain.study.entity.QStudyAttendance;
 import umc.product.domain.study.entity.QStudyMember;
