+++ src/main/java/umc/product/domain/study/service/StudyAttendanceCommandService.java
@@ -1,7 +1,7 @@
 package umc.product.domain.study.service;
 
-import umc.product.domain.study.dto.request.StudyAttendanceRequest;
-import umc.product.domain.study.dto.response.StudyCommonResponse;
+import umc.product.domain.study.dto.common.request.StudyAttendanceRequest;
+import umc.product.domain.study.dto.common.response.StudyCommonResponse;
 import umc.product.domain.study.entity.StudyAttendance;
 
 public interface StudyAttendanceCommandService {
