+++ src/main/java/umc/product/domain/study/service/StudyCommandServiceImpl.java
@@ -4,8 +4,8 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import umc.product.domain.study.dto.request.StudyModifyRequest;
-import umc.product.domain.study.dto.response.StudyCommonResponse;
+import umc.product.domain.study.dto.common.request.StudyModifyRequest;
+import umc.product.domain.study.dto.common.response.StudyCommonResponse;
 import umc.product.domain.study.entity.Study;
 import umc.product.domain.study.entity.StudyMember;
 import umc.product.domain.study.entity.enums.StudyRole;
