+++ src/main/java/umc/product/domain/study/service/StudyQueryService.java
@@ -1,9 +1,9 @@
 package umc.product.domain.study.service;
 
 import umc.product.domain.roadmap.entity.Roadmap;
-import umc.product.domain.study.dto.response.StudyResponse;
-import umc.product.domain.study.dto.response.StudyWeekChecklistResponse;
-import umc.product.domain.study.dto.response.StudyWorkbookResponse;
+import umc.product.domain.study.dto.common.response.StudyResponse;
+import umc.product.domain.study.dto.common.response.StudyWeekChecklistResponse;
+import umc.product.domain.study.dto.common.response.StudyWorkbookResponse;
 import umc.product.domain.study.entity.Study;
 import umc.product.domain.study.entity.StudyMember;
