+++ src/main/java/umc/product/domain/study/service/StudyQueryServiceImpl.java
@@ -5,10 +5,10 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import umc.product.domain.roadmap.entity.Roadmap;
-import umc.product.domain.study.dto.response.StudyMemberResponse;
-import umc.product.domain.study.dto.response.StudyResponse;
-import umc.product.domain.study.dto.response.StudyWeekChecklistResponse;
-import umc.product.domain.study.dto.response.StudyWorkbookResponse;
+import umc.product.domain.study.dto.common.response.StudyMemberResponse;
+import umc.product.domain.study.dto.common.response.StudyResponse;
+import umc.product.domain.study.dto.common.response.StudyWeekChecklistResponse;
+import umc.product.domain.study.dto.common.response.StudyWorkbookResponse;
 import umc.product.domain.study.entity.Study;
 import umc.product.domain.study.entity.StudyMember;
 import umc.product.domain.study.mapper.StudyMapper;
